Correction : toutes les pages index (1 seul form au lieu d'autant par boutons modif et suppr)

+++ src/CaroGFaimBundle/Controller/categorieController.php
@@ -105,7 +105,7 @@ class categorieController extends Controller
         $form = $this->createDeleteForm($categorie);
         $form->handleRequest($request);
 
-        //if ($form->isSubmitted() && $form->isValid())
+        if ($form->isSubmitted() && $form->isValid())
         {
             $em = $this->getDoctrine()->getManager();
             $em->remove($categorie);
